changed pumps files to sensors, updated org grid to change state between view/delete org on button click and styled the header buttons for the org grid
